Plans 2023: Migrate upgrade click handler to reusable hook form
